eval_kmeans_check.py: fix bfloat16->numpy crash in run_kmeans
net.predict() returns yhat in the model's own dtype (bfloat16, from
net.to(torch.bfloat16) in main()); torch.cat(...).numpy() on a
bfloat16 tensor raises 'Got unsupported ScalarType BFloat16' since
numpy has no bfloat16 type. labels already had a .float() cast (line
122); preds (yhat) didn't. Hit on the real cluster run -- the local
smoke test never got far enough into run_kmeans to catch this before
being cancelled.

+++ scripts/eval_kmeans_check.py
@@ -117,7 +117,7 @@ def run_kmeans(net, tasks, pre_dir, config, ctx_size, seed, args, device):
             reduced = gather_along_seq(batch, keep_idx)
 
             preds_by_ctx = net.predict(reduced, [ctx_size], device, task, bool_as_num=ev.bool_as_num)
-            yhat = preds_by_ctx[ctx_size].cpu()
+            yhat = preds_by_ctx[ctx_size].float().cpu()  # bfloat16 has no numpy equivalent
 
             y = (reduced[val_key].squeeze(-1).float()
                  * reduced["is_targets"].to(torch.float32)).sum(dim=1)
